fix(api): report source difficulty as f64
The SourceState.difficulty field was u64, which truncated
fractional difficulties (e.g. 0.001 became 0). Change to
f64 to preserve the value received from the pool.

+++ mujina-miner/src/api_client/types.rs
@@ -97,5 +97,5 @@ pub struct SourceState {
     pub url: Option<String>,
     /// Current share difficulty set by the source.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub difficulty: Option<u64>,
+    pub difficulty: Option<f64>,
 }
